feat: use stored procedure for data insertion and update  models/DAOs/DTOs

+++ infrastructure/entry-points/api-rest/src/main/java/co/com/management/api/dto/models/request/ProductDTO.java
@@ -15,5 +15,5 @@ public class ProductDTO {
     private Integer quantity;
 
     @NotNull
-    private Double unitPrice;
+    private Float unitPrice;
 }
